+++ src/main/java/com/base/community/service/MemberService.java
@@ -9,6 +9,8 @@ import com.base.community.model.entity.MemberSkills;
 import com.base.community.model.repository.MemberRepository;
 import com.base.community.model.repository.MemberSkillsRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -16,8 +18,13 @@ import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.FileCopyUtils;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -28,6 +35,7 @@ import static com.base.community.type.MemberCode.MEMBER_STATUS_ING;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class MemberService implements UserDetailsService {
 
     private final MailComponent mailComponent;
@@ -35,6 +43,13 @@ public class MemberService implements UserDetailsService {
     private final MemberSkillsRepository memberSkillsRepository;
 
     private final PasswordEncoder passwordEncoder;
+
+    @Value("${dir.LocalPath}")
+    String baseLocalPath;
+
+    @Value("${dir.UrlPath}")
+    String baseUrlPath;
+
 
     @Transactional
     public Member signup(SignUpDto signUpDto) {
@@ -226,8 +241,35 @@ public class MemberService implements UserDetailsService {
         memberSkillsRepository.delete(memberSkills);
     }
 
-    public Member uploadProfileImg(Long id, MemberDto form) {
+
+
+    public Member uploadProfileImg(Long id, MultipartFile file) {
+
+        MemberDto form = new MemberDto();
+
+        String saveFilename = "";
+        String urlFilename = "";
+
+        if (file != null) {
+
+            String originalFilename = file.getOriginalFilename();
+            String[] arrFilename = getNewSaveFile(baseLocalPath, baseUrlPath, originalFilename);
+
+            saveFilename = arrFilename[0];
+            urlFilename = arrFilename[1];
+
+            try {
+                File newFile = new File(saveFilename);
+                FileCopyUtils.copy(file.getInputStream(), new FileOutputStream(newFile));
+            } catch (IOException e) {
+                log.info(e.getMessage());
+            }
+        }
+
+        form.setFilename(saveFilename);
+        form.setUrlFilename(urlFilename);
         form.setId(id);
+
         Optional<Member> optionalMember = memberRepository.findById(form.getId());
         if(optionalMember.isEmpty()){
             throw new CustomException(NOT_FOUND_USER);
@@ -239,4 +281,61 @@ public class MemberService implements UserDetailsService {
         return member;
 
     }
+
+    public void deleteMember(Long id) {
+        MemberDto form = new MemberDto();
+        form.setId(id);
+
+        Optional<Member> optionalMember = memberRepository.findById(form.getId());
+        if (optionalMember.isEmpty()) {
+            throw new CustomException(NOT_FOUND_USER);
+        }
+        Member member = optionalMember.get();
+
+        memberRepository.delete(member);
+    }
+
+
+
+
+    private String[] getNewSaveFile(String baseLocalPath, String baseUrlPath, String originalFilename) {
+
+        LocalDate now = LocalDate.now();
+
+        String[] dirs = {
+                String.format("%s/%d/", baseLocalPath, now.getYear()),
+                String.format("%s/%d/%02d/", baseLocalPath, now.getYear(), now.getMonthValue()),
+                String.format("%s/%d/%02d/%02d/", baseLocalPath, now.getYear(), now.getMonthValue(), now.getDayOfMonth())};
+
+        String urlDir = String.format("%s/%d/%02d/%02d/", baseUrlPath, now.getYear(), now.getMonthValue(), now.getDayOfMonth());
+
+        for (String dir : dirs) {
+            File file = new File(dir);
+            if (!file.isDirectory()) {
+                file.mkdir();
+            }
+        }
+
+        String fileExtension = "";
+        if (originalFilename != null) {
+            int dotPos = originalFilename.lastIndexOf(".");
+            if (dotPos > -1) {
+                fileExtension = originalFilename.substring(dotPos + 1);
+            }
+        }
+
+        String uuid = UUID.randomUUID().toString().replaceAll("-", "");
+        String newFilename = String.format("%s%s", dirs[2], uuid);
+        String newUrlFilename = String.format("%s%s", urlDir, uuid);
+        if (fileExtension.length() > 0) {
+            newFilename += "." + fileExtension;
+            newUrlFilename += "." + fileExtension;
+        }
+
+        return new String[]{newFilename, newUrlFilename};
+    }
+
+
+
+
 }
