+++ src/main/resources/application.yml
@@ -11,7 +11,7 @@ spring:
 
   jpa:
     hibernate:
-      ddl-auto: create
+      ddl-auto: update
     show-sql: true
 
   mail:
@@ -27,3 +27,12 @@ spring:
 
   jwt:
     secret: secret
+
+  servlet:
+    multipart:
+      max-file-size: 1MB
+
+
+dir:
+  UrlPath: /files
+  LocalPath: path
